Move MetricQueryType
