+++ src/frontend/packages/cloud-foundry/src/actions/cf-metrics.actions.ts
@@ -1,5 +1,5 @@
-import { MetricQueryType } from '../../../core/src/shared/services/metrics-range-selector.types';
 import { MetricQueryConfig, MetricsAction, MetricsChartAction } from '../../../store/src/actions/metrics.actions';
+import { MetricQueryType } from '../../../store/src/types/metric.types';
 import { PaginatedAction } from '../../../store/src/types/pagination.types';
 import { CF_ENDPOINT_TYPE } from '../cf-types';
