+++ src/frontend/packages/store/src/types/metric.types.ts
@@ -22,3 +22,8 @@ export interface IMetricCell {
   job: string;
   origin: string;
 }
+
+export enum MetricQueryType {
+  QUERY = 'query',
+  RANGE_QUERY = 'query_range'
+}
